Fix some bugs on sql

+++ src/web-apis/LetPortal.Portal/Executions/MySQL/MySqlDynamicListQueryDatabase.cs
@@ -132,7 +132,7 @@ namespace LetPortal.Portal.Executions.MySQL
                             }
 
                         }
-                        response.TotalItems = (long)cmd.ExecuteScalar();
+                        response.TotalItems = (int)cmd.ExecuteScalar();
                     }
                 }
                 mysqlDbConnection.Close();
